requirements.txt
pandas==1.5.3
openpyxl==3.0.10
Flask==2.0.1
pylightxl==1.60
Flask-Cors==3.0.10
python-dotenv==1.0.0
Werkzeug==2.2.2
streamlit-agraph==0.0.28
st-ant-tree==0.0.13
streamlit-tree-select==0.0.5